AIUTA-VLM-R1-main/ai_project/aiuta_vlmr1_project/eval_entropy_sweep.py
"""
Apply CoIN's exact NormalizedEntropy technique to VLM-R1's logits.
Sweep tau threshold to find optimal ER.
"""
import json, sys, os, hashlib
import numpy as np
sys.path.insert(0, os.path.expanduser("~/CoIN/idkvqa"))

from datasets import load_dataset
from vqa_evaluator import VQAEvaluator

# Build GT
print("[1/3] Building GT...")
ds = load_dataset("ftaioli/IDKVQA", split="val")
gt = {"choice_label": {"0": "Yes", "1": "No", "2": "I don't know"}, "images": {}}
image_map = {}
reverse_labels = {"Yes": "0", "No": "1", "I don't know": "2"}

for i, row in enumerate(ds):
    sha1 = hashlib.sha1(row["image"].tobytes()).digest().hex()
    nf = f"val_{sha1}.png"
    counts = {}
    answer_keys = []
    for label, count in row["answers"].items():
        key = reverse_labels.get(label, "2")
        counts[key] = str(count)
        for _ in range(count):
            answer_keys.append(key)
    if nf not in gt["images"]:
        gt["images"][nf] = {"result": "accepted", "questions_answers_pairs": []}
    gt["images"][nf]["questions_answers_pairs"].append({
        "question": row["question"], "answers": answer_keys, "counts": counts,
    })
    image_map[i] = (nf, row["question"])

# Load IDKVQA eval results (has token_entropy per sample)
print("[2/3] Loading IDKVQA results with token entropy...")
idkvqa_path = None
for f in sorted(os.listdir("results/logs")):
    if f.startswith("idkvqa_eval") and f.endswith(".out"):
        idkvqa_path = f"results/logs/{f}"

# Check if we have the detailed JSON
idkvqa_json = None

# Direct load
try:
    with open("results/idkvqa/vlmr1_entropy.json") as _f:
        idkvqa_json = json.load(_f)
        print(f"  Loaded: results/idkvqa/vlmr1_entropy.json")
except: pass
for root, dirs, files in os.walk("results"):
    for f in files:
        if "idkvqa" in f and f.endswith(".json"):
            path = os.path.join(root, f)
            with open(path) as fh:
                data = json.load(fh)
            if "per_sample" in data:
                idkvqa_json = data
                print(f"  Found: {path}")
                break

if idkvqa_json is None:
    print("  No IDKVQA JSON with per_sample found. Need to re-run with token entropy.")
    print("  Falling back to KG benchmark results (no token entropy)...")

    # Use KG benchmark results instead
    with open("results/kg_comparison/benchmark_16843.json") as f:
        bench = json.load(f)
    samples = bench["per_sample"]

    # No token entropy available - simulate with hedging as proxy
    has_entropy = False
else:
    samples = idkvqa_json["per_sample"]
    has_entropy = any("token_entropy" in s for s in samples)
    print(f"  Samples: {len(samples)}, has token_entropy: {has_entropy}")

evaluator = VQAEvaluator(gt, cost=1, log=False)

# Baseline
pairs_base = [(image_map[s["idx"]][0], image_map[s["idx"]][1], s["predicted" if "predicted" in s else "raw_predicted"]) for s in samples]
er_base = evaluator.model_get_effective_reliability(pairs_base)
print(f"\n  Raw VLM-R1 baseline: ER = {er_base}%")

if has_entropy:
    print("\n[3/3] Sweeping entropy threshold (like CoIN NormEntropy)...")
    best_tau = 0
    best_er = er_base
    all_taus = []

    for tau_int in range(0, 100):
        tau = tau_int / 100.0
        pairs = []
        for s in samples:
            fn, q = image_map[s["idx"]]
            entropy = s.get("token_entropy", -1)
            answer = s.get("predicted", s.get("raw_predicted", "I don't know"))

            if entropy >= 0 and entropy > tau:
                pairs.append((fn, q, "I don't know"))
            else:
                pairs.append((fn, q, answer))

        er = evaluator.model_get_effective_reliability(pairs)
        all_taus.append((tau, er))

        if er > best_er:
            best_er = er
            best_tau = tau

    # Print top 10 thresholds
    all_taus.sort(key=lambda x: -x[1])
    print(f"\n  Top 10 tau values:")
    for tau, er in all_taus[:10]:
        marker = " <-- BEST" if tau == best_tau else ""
        print(f"    tau={tau:.2f}  ER={er}%{marker}")

    # Count overrides at best tau
    n_override = sum(1 for s in samples if s.get("token_entropy", -1) > best_tau)

    print(f"\n{'='*70}")
    print(f"FINAL COMPARISON (CoIN exact formula, cost=1)")
    print(f"{'='*70}")
    print(f"  Original LLaVA (paper)               ER =  4.83%")
    print(f"  LLaVA + LP (paper)                   ER = 14.01%")
    print(f"  Our Raw VLM-R1                       ER = {er_base}%")
    print(f"  LLaVA + MaxProb (paper)              ER = 15.94%")
    print(f"  LLaVA + EnergyScore (paper)          ER = 20.45%")
    print(f"  LLaVA + NormEntropy (paper)          ER = 21.12%")
    print(f"  ---")
    print(f"  Our VLM-R1 + NormEntropy(tau={best_tau:.2f})   ER = {best_er}%  ({n_override}/502 -> IDK)")
    print(f"{'='*70}")

else:
    print("\n[3/3] No token entropy available. Re-running IDKVQA with entropy...")
    print("  Run: sbatch slurm/run_idkvqa.sbatch")
    print("  (The existing idkvqa_eval.py already computes token_entropy)")

    # Check if the existing results have it
    for root, dirs, files in os.walk("results"):
        for f in files:
            if f.endswith(".json"):
                path = os.path.join(root, f)
                try:
                    with open(path) as fh:
                        d = json.load(fh)
                    if "per_sample" in d and d["per_sample"]:
                        keys = list(d["per_sample"][0].keys())
                        if "token_entropy" in keys:
                            print(f"  FOUND token_entropy in: {path}")
                except:
                    pass
